GamePanel and ScoreboardPanel update

+++ src/main/java/it/unibo/vampireio/controller/ScoreData.java
@@ -2,13 +2,13 @@ package it.unibo.vampireio.controller;
 
 public class ScoreData {
     private String characterName;
-    private double sessionTime;
+    private long sessionTime;
     private int killCounter;
     private int levelCounter;
     private int coinCounter;
     private int score;
 
-    public ScoreData(String characterName, double sessionTime, int killCounter, int levelCounter, int coinCounter, int score) {
+    public ScoreData(String characterName, long sessionTime, int killCounter, int levelCounter, int coinCounter, int score) {
         this.characterName = characterName;
         this.sessionTime = sessionTime;
         this.killCounter = killCounter;
@@ -29,7 +29,7 @@ public class ScoreData {
         return this.levelCounter;
     }
 
-    public double getSessionTime() {
+    public long getSessionTime() {
         return this.sessionTime;
     }
